+++ R/class_Timeseries.R
@@ -18,6 +18,7 @@ setClassUnion("DateNumber", c("POSIXct", "POSIXt", "Date", 'numeric'))
 #' - `en`: English (default)
 #' - `es`: Spanish
 #' - `fr`: French
+#' - `pt`: Portuguese
 #'
 #' ## Note
 #' Character strings in `Code`, `Label`, and `Description` must all be same length