+++ NEWS.md
@@ -1,12 +1,15 @@
 
 # Slick 1.0.1
 
+- Added Portuguese to the list of supported language. Many thanks to Rodrigo Sant'Ana for translating the text!
 - Added a `Misc` slot to all plot objects. 
 - Added option to show `mean` or `median` in the time series plots shown in the 
   Slick app.
 - Fix issue with subsetting plots by Performance Indicators
 - Modify `plotTimeseries` so it can handle plots with no historical values
 - Fix to `plotQuilt(kable=TRUE)` for identical PI values across MPs
+
+
 
 # Slick 1.0.0